Cleaning up not needed validation Separating classes in different files
